Fix VanillaTransformer forward: convert shift_labels to Long for cross_entropy
Similar to the fix applied to evaluate function in 758d07a, the labels
tensor needs to be converted to Long dtype before passing to F.cross_entropy
which requires int64 targets.

+++ baseline_transformer.py
@@ -214,7 +214,7 @@ class VanillaTransformer(nn.Module):
         # Loss calculation with CLM shift (compatible with DAWN train.py)
         if labels is not None:
             shift_logits = logits[:, :-1, :].contiguous()
-            shift_labels = labels[:, 1:].contiguous()
+            shift_labels = labels[:, 1:].contiguous().long()
             loss = F.cross_entropy(
                 shift_logits.view(-1, self.vocab_size),
                 shift_labels.view(-1),
